refactor api to modules catalog and namespaces
