+++ .eslintignore
@@ -1,3 +1,3 @@
 core/build/*.js
 node_modules
-core/api/**/test
+core/modules/**/test
