+++ core/components/blocks/Header/MicrocartIcon.js
@@ -1,5 +1,5 @@
 import { mapGetters } from 'vuex'
-import { openMicrocart, isMicrocartOpen } from '@vue-storefront/core/api/cart'
+import { openMicrocart, isMicrocartOpen } from '@vue-storefront/core/modules/cart'
 
 export default {
   name: 'MicrocartIcon',
