+++ core/components/blocks/Header/WishlistIcon.js
@@ -1,4 +1,4 @@
-import { openWishlist, closeWishlist, isWishlistOpen } from '@vue-storefront/core/api/wishlist'
+import { openWishlist, closeWishlist, isWishlistOpen } from '@vue-storefront/core/modules/wishlist'
 
 export default {
   name: 'WishlistIcon',
