+++ core/components/blocks/Microcart/Microcart.js
@@ -3,7 +3,7 @@ import { mapGetters } from 'vuex'
 import VueOfflineMixin from 'vue-offline/mixin'
 
 // Core dependecies
-import { productsInCart, closeMicrocart, isMicrocartOpen, removeFromCart } from '@vue-storefront/core/api/cart'
+import { productsInCart, closeMicrocart, isMicrocartOpen, removeFromCart } from '@vue-storefront/core/modules/cart'
 
 // Core mixins
 import onEscapePress from '@vue-storefront/core/mixins/onEscapePress'
