+++ core/components/blocks/Wishlist/Wishlist.js
@@ -1,5 +1,5 @@
 import onEscapePress from '@vue-storefront/core/mixins/onEscapePress'
-import { loadWishlist, productsInWishlist, closeWishlist, isWishlistOpen } from '@vue-storefront/core/api/wishlist'
+import { loadWishlist, productsInWishlist, closeWishlist, isWishlistOpen } from '@vue-storefront/core/modules/wishlist'
 
 export default {
   name: 'Wishlist',
