+++ core/modules/offline-order/helpers/onNetworkStatusChange.ts
