+++ core/modules/user/logOut.js
