+++ core/modules/wishlist/clearWishlist.js
