+++ core/modules/wishlist/features/closeWishlist.js
