+++ core/modules/wishlist/features/isWishlistOpen.js
