+++ core/modules/wishlist/index.js
