+++ core/modules/wishlist/removeFromWishlist.js
