+++ doc/api-modules/cart.md
@@ -35,7 +35,7 @@ import {
   closeMicrocart,
   openMicrocart,
   isMicrocartOpen
-} from '@vue-storefront/core/api/cart'
+} from '@vue-storefront/core/modules/cart'
 
 export default {
   //...other properties
