+++ tests/unit/core/api/wishlist/TestInstanceForLoadTest.vue
@@ -4,7 +4,7 @@
 
 <script>
 // Import the features you need in the view grouped in modules
-import { loadWishlist } from '@vue-storefront/core/api/wishlist'
+import { loadWishlist } from '@vue-storefront/core/modules/wishlist'
 
 export default {
   mixins: [
